Review 3rd week

고생하셨습니다.

+++ backend/src/main/java/com/team6/issue_tracker/domain/page/dto/IssuePageResponse.java
@@ -2,7 +2,7 @@ package com.team6.issue_tracker.domain.page.dto;
 
 import com.team6.issue_tracker.domain.label.dto.LabelDto;
 import com.team6.issue_tracker.domain.member.dto.MemberDto;
-import com.team6.issue_tracker.domain.milestone.Milestone;
+import com.team6.issue_tracker.domain.milestone.domain.Milestone;
 import lombok.Builder;
 import lombok.Data;
 
@@ -10,8 +10,9 @@ import java.util.List;
 
 @Data @Builder
 public class IssuePageResponse {
-    private Integer openIssueCount;
-    private Integer closedIssueCount;
+
+    private Long openIssueCount;
+    private Long closedIssueCount;
     private Integer page;
     private Integer openIssueMaxPage;
     private Integer closeIssueMaxPage;
